requirements.txt
pandas
numpy
sqlalchemy
pyodbc
python-dotenv
telethon
aiohttp
beautifulsoup4
